build: change how we build the project
