+++ .npmrc
@@ -1,0 +1,1 @@
+ignore-workspace-root-check=true
