+++ package.json
@@ -1,38 +1,75 @@
 {
   "name": "time2fa",
+  "type": "module",
   "version": "1.2.0",
   "description": "NodeJS OTP implementation",
-  "repository": {
-    "type": "git",
-    "url": "git+ssh://git@github.com/PlanetHoster/time2fa.git"
-  },
-  "main": "./main.js",
-  "typings": "./main.d.ts",
-  "scripts": {
-    "build": "npm i && npm run clean && tsc && node copyBuildFiles.js",
-    "test": "npm run build && jest",
-    "clean": "rimraf ./.build",
-    "lint": "eslint src",
-    "format": "npx prettier --write src"
-  },
   "author": "Marc-Andre Godin",
   "license": "MIT",
+  "homepage": "https://github.com/PlanetHoster/time2fa#readme",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/PlanetHoster/time2fa.git"
+  },
+  "bugs": "https://github.com/PlanetHoster/time2fa/issues",
+  "sideEffects": false,
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "require": "./dist/index.cjs",
+      "import": "./dist/index.mjs"
+    }
+  },
+  "main": "./dist/index.mjs",
+  "module": "./dist/index.mjs",
+  "types": "./dist/index.d.ts",
+  "typesVersions": {
+    "*": {
+      "*": [
+        "./dist/*",
+        "./dist/index.d.ts"
+      ]
+    }
+  },
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "unbuild",
+    "dev": "unbuild --stub",
+    "lint": "eslint src",
+    "lint-staged": "lint-staged",
+    "prepublishOnly": "nr build",
+    "release": "bumpp && npm publish",
+    "start": "esno src/index.ts",
+    "test": "vitest run --coverage",
+    "typecheck": "tsc --noEmit"
+  },
   "devDependencies": {
-    "@types/base-x": "^3.0.6",
-    "@types/jest": "^29.5.0",
+    "@antfu/eslint-config": "^0.38.4",
+    "@antfu/ni": "^0.21.2",
     "@types/node": "^18.15.11",
-    "@typescript-eslint/eslint-plugin": "^5.57.0",
-    "@typescript-eslint/parser": "^5.57.0",
-    "eslint": "^8.37.0",
+    "@typescript-eslint/eslint-plugin": "^5.59.0",
+    "@typescript-eslint/parser": "^5.59.0",
+    "@vitest/coverage-c8": "^0.30.1",
+    "bumpp": "^9.1.0",
+    "eslint": "^8.38.0",
     "eslint-config-prettier": "^8.8.0",
+    "esno": "^0.16.3",
+    "lint-staged": "^13.2.0",
     "prettier": "^2.8.7",
-    "ts-jest": "^29.1.0",
-    "typescript": "^5.0.2"
+    "rimraf": "^5.0.0",
+    "simple-git-hooks": "^2.8.1",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.0.3",
+    "unbuild": "^1.2.0",
+    "vitest": "^0.30.1"
   },
-  "bugs": {
-    "url": "https://github.com/PlanetHoster/time2fa/issues"
+  "simple-git-hooks": {
+    "pre-commit": "npm run lint-staged"
   },
-  "homepage": "https://github.com/PlanetHoster/time2fa/blob/main/README.md",
+  "lint-staged": {
+    "src/*": "eslint --fix"
+  },
   "keywords": [
     "otp",
     "totp",
